Implement Frozen as an alternative to ALBD > 0.4
The HEMCO emissions extensions sometimes use a check of albedo > 0.4 to
determine if ground is covered in snow or ice. This seems indirect, and
can cause issues in (eg) desert terrain. Instead, we now expect an
external state variable called "FROZEN" to be set. If FROZEN is greater
than 0.5, the ground is considered to be covered in snow or ice. Users
can exactly reproduce the old behavior by setting something like

Frozen = where(Albedo > 0.4, 1.0, 0.0)

Signed-off-by: Sebastian D. Eastham <seastham@mit.edu>